+++ tests/ExampleGenerationController/Element/GetIndexTest.php
@@ -32,6 +32,20 @@ class GetIndexTest extends BaseRequestTestCase
         );
     }
 
+    public function testGetIndexRedirect304(): void
+    {
+        $response = $this->runGetRequest('/', self::TOKEN);
+        $etag = $response->getHeader('ETag');
+        $response = $this->runGetRequest('/', self::TOKEN, ['If-None-Match' => $etag]);
+        $this->assertNotModifiedResponse($response);
+        $documentationHeadersPath = 'docs/api-endpoints/element/get-index/304-response-header.txt';
+        $this->assertHeadersInDocumentationAreIdenticalToHeadersFromRequest(
+            self::PATH_TO_ROOT,
+            $documentationHeadersPath,
+            $response
+        );
+    }
+
     public function testGetIndexFailure401(): void
     {
         $response = $this->runGetRequest('/', 'thisTokenDoesNotExist');