+++ tests/FeatureTests/BaseRequestTestCase.php
@@ -272,6 +272,13 @@ abstract class BaseRequestTestCase extends TestCase
         $this->assertFalse($response->hasHeader('Location'));
     }
 
+    public function assertNotModifiedResponse(ResponseInterface $response): void
+    {
+        $this->assertSame(304, $response->getStatusCode());
+        $this->assertEmpty((string) $response->getBody());
+        $this->assertFalse($response->hasHeader('Location'));
+    }
+
     public function assertIsDeletedResponse(ResponseInterface $response): void
     {
         $this->assertSame(204, $response->getStatusCode());